LayoutTests/imported/w3c/web-platform-tests/shadow-dom/focus-navigation/focus-navigation-slot-nested-2levels-expected.txt


FAIL Focus controller should treat each slot as a focus scope. assert_equals: Focus should move forward from nested1/inner-before to nested1/button expected Element node <button id="button"><slot></slot></button> but got Element node <input id="innermost-before"></input>


button
